refactor: split model binding helpers
